1) Add reusable recursion for the roll up feature 2) Remove files
