+++ packages/zkdb/src/rollup/index.ts
@@ -1,5 +1,4 @@
 export * from './action.js'
-export * from './offchain-rollup.js';
+export * from './rollup-program.js';
 export * from './rollup-params.js';
 export * from './rollup-service.js'
-export * from './rollup-contract.js'
